requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
click==7.1.2
Django==3.1.6
djangorestframework==3.12.2
flake8==3.8.4
gunicorn==20.0.4
importlib-metadata==3.4.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2021.1
regex==2020.11.13
schedule==1.0.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
whitenoise==5.2.0
zipp==3.4.0
